feat: cli can create config file on demand
